Use resilient HttpClient everywhere
