+++ src/NexusMods.Networking.HttpDownloader/Services.cs
@@ -1,5 +1,8 @@
+using System.Net;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Http.Resilience;
 using NexusMods.App.BuildInfo;
+using Polly;
 
 namespace NexusMods.Networking.HttpDownloader;
 
@@ -12,9 +15,30 @@ public static class Services
     {
         return services.AddSingleton<HttpClient>(_ =>
         {
-            var client = new HttpClient();
-            client.DefaultRequestHeaders.UserAgent.ParseAdd(ApplicationConstants.UserAgent);
+            var client = BuildClient();
             return client;
         });
+    }
+
+    private static HttpClient BuildClient()
+    {
+        var pipeline = new ResiliencePipelineBuilder<HttpResponseMessage>()
+            .AddRetry(new HttpRetryStrategyOptions())
+            .Build();
+
+        HttpMessageHandler handler = new ResilienceHandler(pipeline)
+        {
+            InnerHandler = new SocketsHttpHandler(),
+        };
+
+        var client = new HttpClient(handler)
+        {
+            DefaultRequestVersion = HttpVersion.Version11,
+            DefaultVersionPolicy = HttpVersionPolicy.RequestVersionOrHigher,
+        };
+
+        client.DefaultRequestHeaders.UserAgent.ParseAdd(ApplicationConstants.UserAgent);
+
+        return client;
     }
 }
